+++ CHANGELOG.md
@@ -20,6 +20,7 @@ Models
 Extensions
 
 - `RectangleCornerRadii.init(_:)` initializer is added, that initializes `RectangleCornerRadii` with value
+- `View.dynamicTypeSize(type:)` method is added, that sets `DynamicTypeSize` based on `DynamicTypeSizeType`.
 - `View.cornerRadius(_:corners:)` has been removed, as it's obsoleted by `UnevenRoundedRectangle`
 - `View.cornerRadius(_:uiCorners:)` has been removed, as it's obsoleted by `UnevenRoundedRectangle`